TST: merge mass balance/single phase flow
